Stop using configure-aspen.py in http://aspen.io/
